+++ src/main/java/com/timeflowsystem/security/filter/JWTTokenGeneratorFilter.java
@@ -1,0 +1,60 @@
+package com.timeflowsystem.security.filter;
+
+import com.timeflowsystem.security.constants.ApplicationConstants;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.core.env.Environment;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.crypto.SecretKey;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
+import java.util.stream.Collectors;
+
+public class JWTTokenGeneratorFilter extends OncePerRequestFilter {
+
+    /**
+     * @param request
+     * @param response
+     * @param filterChain
+     * @throws ServletException
+     * @throws IOException
+     */
+    @Override
+    protected void doFilterInternal(HttpServletRequest request,
+                                    HttpServletResponse response,
+                                    FilterChain filterChain) throws ServletException, IOException {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (null != authentication) {
+            Environment env = getEnvironment();
+            if (null != env) {
+                String secret = env.getProperty(ApplicationConstants.JWT_SECRET_KEY, // If it's not informed JWT_SECRET_KEY
+                        ApplicationConstants.JWT_SECRET_DEFAULT_VALUE);             // use JWT_SECRET_DEFAULT_VALUE
+                SecretKey secretKey = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
+                String jwt = Jwts.builder().issuer("Timeflow").subject("JWT Token")
+                        .claim("username", authentication.getName())
+                        .claim("authorities", authentication.getAuthorities().stream().map(
+                                GrantedAuthority::getAuthority).collect(Collectors.joining(",")))
+                        .issuedAt(new Date())
+                        .expiration(new Date((new Date()).getTime() + 30000000)) //current date + (30000000 milisseconds)
+                        .signWith(secretKey).compact();
+                response.setHeader(ApplicationConstants.JWT_HEADER, jwt);
+            }
+        }
+        filterChain.doFilter(request, response); // Call next filter in the chain
+    }
+
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException {
+        return !request.getServletPath().equals("/user"); // Only execute the filter for this endpoint
+    }
+
+}